Qualify for the "evilphk" monicker:
For a limited number of cases we can now esi:include a gzip'ed
object into a gzip'ed ESI object.

There is less than 1 chance in hell this will work for you in
practice, I need to handle the N-1 other cases of last bit
and stop bit combinations in the gziped object, but that is
just a small (but tedious) matter of programming.