The CLI protocol and its implementation is not designed for transferring huge amounts of data, and it should not be used that way.
But we have commands like ban.list which potentially can return at
metric shitload of data and this is a problem.

This commit implements a cli_limit paramter which (with its upper
limit) protects from responses that cannot be moved by the CLI
protocol no matter what.

We set the default to a low 4096 bytes, in order to make people
specifically shoot their feet off if they really want to see 200k
bans.