Add a new VTLA:  VSA.
This does absolutely not stand for "suckaddr", and but it tries to
sanitize the sockaddr_* mess enough that we can simply pass around
struct sockaddr_storage without also tracking its length.

The cost is that we do know something about address families (IPv4, IPv6),
which strictly speaking should not be necessary, had the sockaddr
not been messed up by the IPv6 people.