Don't close and reopen(+bind) the acceptor socket in the server instances whenever we give the server instance more work, doing so runs into a dark and murky corner somewhere in Stevens and Solaris sometimes doesn't like that.
The trouble is that the initial bind(2) asks the kernel to assign
a port, but the subsequent insist on reusing that port, and that,
in some cases, could mean toruble which cannot be allowed.

Add a "-break" instruction to server so that test case c00035 can
still work.

Spotted by:	Nils Goroll / UPLEX