A `cmd` feature for custom shell-based checks
In addition to harcoded features in varnishtest, this opens a window for
out-of-tree uses of the test framework to skip test cases if an external
component (a database system, OS-specific capabilities, an environment
variable, a library, etc) is missing.

This feature takes an extra argument, a command-line that must exit with
a zero status. Complex feature testing can nicely be wrapped in scripts
at the user's discretion:

    feature cmd "my --command=line"

If the test is skipped, it is logged as:

    **   top   0.0 === feature cmd "my --command=line"
    *    top   0.0 SKIPPING test, missing feature: my --command=line

If the command-line is missing, it is logged as:

    **   top   0.0 === feature cmd
    ---- top   0.0 Missing the command line