Fix clang-specific warnings
This round of fixes was originally from work-in-progress in https://github.com/varnishcache/varnish-cache/issues/2174 but
the way forward is to remove the autotools. Since https://github.com/varnishcache/varnish-cache/issues/2174 became obsolete
they should be checked in independently.

Those are warnings we should be able to comply to, but are commented
out since they are clang-specific, but I'm not sure how `optreset`
should be handled so I left that one alone.

Closes https://github.com/varnishcache/varnish-cache/issues/2174