typedefs for real/mono time and durations
We use double for all time representations, yet monotonic time,
real time and durations are not to be confused.

Also, we might want to change the representation of time in the
future.

To get an implicit documentation of the semantic type of a time
value and to facilitate working on the latter, we start off by
introducing simple typedefs which neither inply any change nor
offer any type checking. But they pave the way...