Add a hint in the accept_filter documentation
Trying to understand why I was seeing setsockopt fail on FreeBSD VTEST
boxes and why it failed with an undocumented ENOENT I finally realized
that I needed to kldload accf_http beforehand.

The ENOENT errno was probably a result of accept_filter code not finding
a kernel module for "httpready".