varnishtest logexpect: add code for negative matching conditions
We add a list of expects which, if matched, will trigger failure.

When we encounter an expect with skip_max == LE_FAIL, we put it onto the
"fail list". For skip_max == LE_CLEAR, we clear that list.

For every VSL line, all expects on the fail list are tested, if any.

When the list of (positive) expects has been completed, logexpect
continues testing the fail list until the transaction is complete. This
obviously does not work for raw grouping, so in this case the test
author needs to ensure that the fail list be empty when the logexpect is
expected to complete.

Ref https://github.com/varnishcache/varnish-cache/issues/3249