vre: Extract VRE_error() from VRE_compile()
This is a major step back in terms of error reporting, but I haven't
found anything in libpcre to translate an error code to an error
message. "pcre error %d" should still be a good enough hint.

Looking forward, we are going to need this for libpcre2 but again the
API leaves some to be desired and only works by writing error messages
to buffers.

The return value implies that we got a valid error code, which will be
verified with pcre2. I couldn't find how to verify error codes with pcre.
Writing to the VSB is however fail-safe, it's the caller's job to deal
with a VSB error.