Check for the empty abstract socket name
As pointed out by Dridi, we should rather not make an attempt to
support un-printable socket names, of which the empty name is
the most prominent case.

For all other non-printable cases (e.g. d\0r\0i\0d\0i), we have
no support to pass them in the first place, because we treat
uds paths as NUL-terminated strings.