fix bug in abstract sockets
with uds abstract sockets, sun_path should start with a NULL character followed
by the socket's name. The name is not considered to be NULL terminated and can
contain NULL bytes which have no special meaning. socklen is used to determine
the length of name and must be set to the length of the struct sockaddr_un up to
the last character of name, otherwise the 108 characters of sun_path will be
treated as the name of the socket, including NULL bytes.