Makefile: remove target files on python failure
When running the tool on a python-less machine, it leaves the last
python attempt's output in the target (empty) file. This causes any
subsequent calls to try to continue as if the file creation had
succeeded instead of attempting to build the files again, making it
more difficult to spot that the issue is python missing from the
machine. Let's just remove the output file on failure.

This can easily be emulated with the following command twice:

  $ make PYTHON=/bin/false PYTHON2=/bin/false PYTHON3=/bin/false
  trying python3
  trying python2 instead
  trying python instead
  failed
  make: *** [Makefile:67: src/vtc_h2_dectbl.h] Error 1

  $ make PYTHON=/bin/false PYTHON2=/bin/false PYTHON3=/bin/false
  cc \
          -O2 -s -Wall -Werror \
          -o vtest \
          -Isrc -Ilib -I/usr/local/include -pthread \
          src/*.c lib/*.c \
          -L/usr/local/lib -lm -lpcre2-8 -lz
  src/vtc_h2_hpack.c: In function 'huff_decode':
  src/vtc_h2_hpack.c:66:29: error: 'tbl_0' undeclared (first use in this function); did you mean 'tbl'?
     66 |         struct stbl *tbl = &tbl_0;
        |                             ^~~~~
        |                             tbl
  src/vtc_h2_hpack.c:66:29: note: each undeclared identifier is reported only once for each function it appears in
  (...)

Now it will stick to the first failure when repeated.