Start making vtest command arguments less positional
Our common pattern of argument parsing is basically

	for (; *av != NULL; av++) {
		// parse first set of arguments
		if (unknown_argument)
			break;
	}
	for (; *av != NULL; av++) {
		// parse second set of arguments
		if (unknown_argument)
			break;
	}
	// ...

This makes vtest command arguments "semi-positional": Within each set,
they behave like named arguments, but once parsing progressed to the
next set, they behave like positional arguments. This is explicitly
mentioned in the documentation, for example for txreq/txresp:

	"These three switches can appear in any order but must come
	 before the following ones."

While properly documented, this is un-POLA behavior.

This patch changes this behavior for the first set of txreq/txresp
arguments in a simplistic way by filtering known arguments out of the
argument vector. The drawback of the simplistic approach is that
argument values of a "later stage" can not be the same string as
argument names of an "earlier stage". For example, the following would
no onger work:

	txreq -body -nouseragent